Object Pool, Tower Aim and Currency
Made an object pool for only a certain amount of rams may spawn into the game. Added A way for the tower to aim at the closest enemy and at a certain range. Added a currency system where destroying a ram or letting them trough will add or deduct currency.