UI, Bug Fixes, and Lose Condition
Add a the amount of gold the player has onto the screen. FIxed and issue where the tower would keep firing when there are no rams on the field. When the player let the enemy pass by, they lose gold. When gold reaches below 0, they lose and reload the scene.